refactor: apply email verification only to customers
