+++ backend/src/routes/usersRouter.ts
@@ -3,7 +3,6 @@ import {
   authenticateUserController,
   sendUserResetEmailController,
   updatePasswordController,
-  verifyUserEmailController,
 } from "../controllers/usersController";
 
 export const usersRouter = express.Router();
@@ -12,5 +11,4 @@ export const usersRouter = express.Router();
 
 usersRouter.post("/authenticate", authenticateUserController);
 usersRouter.post("/send-password-reset", sendUserResetEmailController);
-usersRouter.patch("/verify-email", verifyUserEmailController);
 usersRouter.patch("/update-password", updatePasswordController);
