+++ frontend/src/app/helper/api/usersApi.ts
@@ -1,11 +1,7 @@
 import {
   CONFIRMATION_EMAIL_RESEND_FAILURE,
   EMAIL_NOT_REGISTERED_MESSAGE,
-  EMAIL_VERIFICATION_FAILED_MESSAGE,
   EMAIL_VERIFICATION_RESENT_NOTICE,
-  EMAIL_VERIFICATION_SUCCESS_MESSAGE,
-  EMAIL_VERIFICATION_TOKEN_EXPIRED,
-  EMAIL_VERIFICATION_UNEXPECTED_ERROR,
   GENERAL_ERROR_MESSAGE,
   GENERAL_ERROR_MESSAGE_JA,
   LOGIN_FAILED_MESSAGE,
@@ -60,48 +56,6 @@ export const authenticateUser = async (
     return {
       errorMessage:
         language === "ja" ? GENERAL_ERROR_MESSAGE_JA : GENERAL_ERROR_MESSAGE,
-    };
-  }
-};
-
-export const verifyUserEmail = async (
-  token: string,
-  userType: UserType,
-): Promise<{ success?: string; error?: string }> => {
-  const apiUrl = `${BACKEND_ORIGIN}/users/verify-email`;
-  const headers = { "Content-Type": "application/json" };
-  const body = JSON.stringify({
-    token,
-    userType,
-  });
-
-  const statusErrorMessages: Record<number, string> = {
-    400: EMAIL_VERIFICATION_FAILED_MESSAGE,
-    404: EMAIL_VERIFICATION_FAILED_MESSAGE,
-    410: EMAIL_VERIFICATION_TOKEN_EXPIRED,
-  };
-
-  try {
-    const response = await fetch(apiUrl, {
-      method: "PATCH",
-      headers,
-      body,
-    });
-
-    const errorMessage = statusErrorMessages[response.status];
-    if (errorMessage) {
-      return { error: errorMessage };
-    }
-
-    if (!response.ok) {
-      throw new Error(`HTTP Status: ${response.status} ${response.statusText}`);
-    }
-
-    return { success: EMAIL_VERIFICATION_SUCCESS_MESSAGE };
-  } catch (error) {
-    console.error("API error while verifying user email:", error);
-    return {
-      error: EMAIL_VERIFICATION_UNEXPECTED_ERROR,
     };
   }
 };
